[KAN-85] controller 분리
