+++ src/hb-backend-api/auth/adapters/in/rest/auth-refresh-token.controller.ts
@@ -1,53 +1,23 @@
-import {
-  Body,
-  Controller,
-  Inject,
-  Post,
-  Req,
-  Res,
-  UseInterceptors,
-} from "@nestjs/common";
-import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
+import { Controller, Inject, Post, Req, Res } from "@nestjs/common";
+import { ApiOperation, ApiTags } from "@nestjs/swagger";
 import { Request, Response } from "express";
-import { LoginAuthUseCase } from "../../../application/ports/in/login-auth.use-case";
 import { EndPointPrefixConstant } from "../../../../../shared/constants/end-point-prefix.constant";
-import { LoginAuthDto } from "../dto/login-auth.dto";
-import { GetLoginTokenDto } from "../dto/get-login-token.dto";
-import { LoginAuthCommand } from "../../../application/command/login-auth.command";
 import { RefreshAuthTokenUseCase } from "../../../application/ports/in/refresh-auth-token.use-case";
-import { CookiesInterceptor } from "../../../../../shared/adpaters/in/rest/interceptors/cookie.interceptor";
 import { DIToken } from "../../../../../shared/di/token.di";
 import { RefreshToken } from "../../../domain/vo/refresh-token.vo";
 
 @ApiTags("Auth")
 @Controller(`${EndPointPrefixConstant}/auth`)
-export class AuthController {
+export class AuthRefreshTokenController {
   private ACCESS_TOKEN = "accessToken";
   private REFRESH_TOKEN = "refreshToken";
   private ACCESS_TOKEN_EXPIRATION = 24 * 60 * 60 * 1000;
   private REFRESH_TOKEN_EXPIRATION = 30 * 24 * 60 * 60 * 1000;
 
   constructor(
-    @Inject(DIToken.AuthModule.LoginAuthUseCase)
-    private readonly loginAuthUseCase: LoginAuthUseCase,
     @Inject(DIToken.AuthModule.RefreshAuthTokenUseCase)
     private readonly refreshTokenUseCase: RefreshAuthTokenUseCase,
   ) {}
-
-  @ApiOperation({
-    summary: "사용자 로그인",
-    description: "사용자 로그인",
-  })
-  @ApiResponse({ type: GetLoginTokenDto })
-  @Post("/login")
-  @UseInterceptors(CookiesInterceptor)
-  public async login(@Body() body: LoginAuthDto): Promise<GetLoginTokenDto> {
-    const loginAuthResult = await this.loginAuthUseCase.invoke(
-      LoginAuthCommand.of(body.nickname, body.password),
-    );
-
-    return GetLoginTokenDto.from(loginAuthResult);
-  }
 
   @ApiOperation({
     summary: "리프래시 토큰 갱신",
