+++ src/hb-backend-api/auth/auth.module.ts
@@ -6,7 +6,7 @@ import { PassportModule } from "@nestjs/passport";
 import { JwtAuthAdapter } from "../../infra/adapters/jwt/jwt-auth.adapter";
 import { AuthEntity } from "./domain/entity/auth.entity";
 import { AuthSchema } from "./domain/entity/auth.schema";
-import { AuthController } from "./adapters/in/rest/auth.controller";
+import { AuthRefreshTokenController } from "./adapters/in/rest/auth-refresh-token.controller";
 import { AuthPersistenceAdapter } from "./adapters/out/persistence/auth-persistence.adapter";
 import { AuthQueryAdapter } from "./adapters/out/query/auth-query.adapter";
 import { LoginAuthService } from "./application/use-cases/login-auth.service";
@@ -15,6 +15,7 @@ import { RefreshTokenAuthService } from "./application/use-cases/refresh-token-a
 import { JwtStrategy } from "../../shared/adpaters/in/rest/strategy/jwt.strategy";
 import { AuthRepositoryImpl } from "./infra/repositories/auth.repository.impl";
 import { DIToken } from "../../shared/di/token.di";
+import { AuthLoginController } from "./adapters/in/rest/auth-login.controller";
 
 @Module({
   imports: [
@@ -67,7 +68,7 @@ import { DIToken } from "../../shared/di/token.di";
       useClass: RefreshTokenAuthService,
     },
   ],
-  controllers: [AuthController],
+  controllers: [AuthLoginController, AuthRefreshTokenController],
   exports: [
     JwtModule,
     MongooseModule,
