+++ src/hb-backend-api/category/category.module.ts
@@ -6,7 +6,7 @@ import { DailyTodoModule } from "../daily-todo/daily-todo.module";
 import { DIToken } from "../../shared/di/token.di";
 import { CategoryRepositoryImpl } from "./infra/repositories/category.repository.impl";
 import { CategoryPersistenceAdapter } from "./adapters/out/persistence/category-persistence.adapter";
-import { CategoryController } from "./adapters/in/rest/category.controller";
+import { DeleteCategoryByIdController } from "./adapters/in/rest/delete-category-by-id.controller";
 import { CreateCategoryService } from "./application/use-cases/create-category.service";
 import { CategoryQueryAdapter } from "./adapters/out/query/category-query.adapter";
 import { UserModule } from "../user/user.module";
@@ -14,6 +14,10 @@ import { GetAllCategoryService } from "./application/use-cases/get-all-category.
 import { PatchCategoryService } from "./application/use-cases/patch-category.service";
 import { GetCategoryService } from "./application/use-cases/get-category.service";
 import { DeleteCategoryService } from "./application/use-cases/delete-category.service";
+import { CreateCategoryController } from "./adapters/in/rest/create-category.controller";
+import { GetAllCategoryController } from "./adapters/in/rest/get-all-category.controller";
+import { GetCategoryByIdController } from "./adapters/in/rest/get-category-by-id.controller";
+import { UpdateCategoryTitleController } from "./adapters/in/rest/update-category-title.controller";
 
 @Module({
   imports: [
@@ -26,7 +30,13 @@ import { DeleteCategoryService } from "./application/use-cases/delete-category.s
     DailyTodoModule,
     UserModule,
   ],
-  controllers: [CategoryController],
+  controllers: [
+    CreateCategoryController,
+    DeleteCategoryByIdController,
+    GetAllCategoryController,
+    GetCategoryByIdController,
+    UpdateCategoryTitleController,
+  ],
   providers: [
     {
       provide: DIToken.CategoryModule.CategoryRepository,
